Make the MAC detection more resilient
The original code only fetched the MAC from the message.
The way I understand it, DHCPv6 requests focus on the DUID as the identifier, NOT the MAC.
This is especially true if the client is requesting a temporary or private IP address.

This change falls back to trying to extract the IP address of the peer we're talking to.
It'll be a link local address which starts with fe80:.

Sometimes, that link-local address is generated as an EUI64 IPv6 address which encodes the device's MAC address.
This isn't universally true: sometimes devices will deliberately not expose an EUI64 address for privacy reasons.
However, we can detect that case as eui64 IPs have 0xFFFE tucked in the middle of the MAC.

Anyway, if we can extract a MAC from the eui64 IP from the peer, we use that as their MAC.

Note that I don't believe this will work if we have Relay Agent forwarding in our network.
We don't right now, so ... that is a thing to watch out for.